changed crawlid's to ints in network measurement update

+++ src/entities/NetworkMeasurementUpdate.ts
@@ -10,9 +10,9 @@ export default class NetworkMeasurementUpdate {
         // @ts-ignore
     id: number;
 
-    @Column("bigint", {default: 0})
+    @Column("int", {default: 0})
     startCrawlId: number = 0;
-    @Column("bigint", {default: 0})
+    @Column("int", {default: 0})
     endCrawlId: number = 0;
 
 }
